Refactor structure of resource tracking
