+++ packages/core/src/rum/instrumentation/resourceTracking/distributedTracing/__tests__/firstPartyHosts.test.ts
