+++ packages/core/src/rum/instrumentation/resourceTracking/distributedTracing/distributedTracingHeaders.ts
