+++ packages/core/src/rum/instrumentation/resourceTracking/requestProxy/XHRProxy/URLHostParser.ts
@@ -4,9 +4,9 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-import { InternalLog } from '../../../../InternalLog';
-import { SdkVerbosity } from '../../../../SdkVerbosity';
-import type { Hostname } from '../domain/firstPartyHosts';
+import { InternalLog } from '../../../../../InternalLog';
+import { SdkVerbosity } from '../../../../../SdkVerbosity';
+import type { Hostname } from '../../distributedTracing/firstPartyHosts';
 
 // matches what is between the first "://" and the next "/", ":" or whitespace
 const hostRegex = '^.+://([^:/\\s]+)';
