+++ packages/core/src/rum/instrumentation/resourceTracking/requestProxy/XHRProxy/__tests__/URLHostParser.test.ts
