+++ packages/core/src/rum/instrumentation/resourceTracking/requestProxy/interfaces/RequestProxy.ts
